Make MDE package pass intel IPF compiler with /W4 /WX switched on.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@2312 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/MdePkg/Library/BaseLib/SwapBytes16.c
@@ -32,5 +32,5 @@ SwapBytes16 (
   IN      UINT16                    Operand
   )
 {
-  return (Operand << 8) | (Operand >> 8);
+  return (UINT16) ((Operand << 8) | (Operand >> 8));
 }
